+++ frontend/constants.js
@@ -9,10 +9,6 @@ const OBJECT_ID = Symbol('_objectId')  // the object ID of the current object (s
 const CONFLICTS = Symbol('_conflicts') // map or list (depending on object type) of conflicts
 const CHANGE    = Symbol('_change')    // the context object on proxy objects used in change callback
 
-// Properties of Automerge list objects
-const ELEM_IDS  = Symbol('_elemIds')   // list containing the element ID of each list element
-const MAX_ELEM  = Symbol('_maxElem')   // maximum element counter value in this list (number)
-
 module.exports = {
-  OPTIONS, CACHE, INBOUND, STATE, OBJECT_ID, CONFLICTS, CHANGE, ELEM_IDS, MAX_ELEM
+  OPTIONS, CACHE, INBOUND, STATE, OBJECT_ID, CONFLICTS, CHANGE
 }