+++ test/frontend_test.js
@@ -93,9 +93,9 @@ describe('Automerge.Frontend', () => {
       const birds = Frontend.getObjectId(doc.birds), actor = Frontend.getActorId(doc)
       assert.deepEqual(doc, {birds: ['chaffinch']})
       assert.deepEqual(req, {requestType: 'change', actor, seq: 1, deps: {}, ops: [
-        {obj: ROOT_ID, action: 'makeList', key: 'birds',      child: birds},
-        {obj: birds,   action: 'ins',      key: '_head',      elem: 1},
-        {obj: birds,   action: 'set',      key: `${actor}:1`, value: 'chaffinch'},
+        {obj: ROOT_ID, action: 'makeList', key: 'birds', child: birds},
+        {obj: birds,   action: 'ins',      key: 0},
+        {obj: birds,   action: 'set',      key: 0, value: 'chaffinch'},
       ]})
     })
 
@@ -106,7 +106,7 @@ describe('Automerge.Frontend', () => {
       assert.deepEqual(doc1, {birds: ['chaffinch']})
       assert.deepEqual(doc2, {birds: ['greenfinch']})
       assert.deepEqual(req2, {requestType: 'change', actor, seq: 2, deps: {}, ops: [
-        {obj: birds, action: 'set', key: `${actor}:1`, value: 'greenfinch'}
+        {obj: birds, action: 'set', key: 0, value: 'greenfinch'}
       ]})
     })
 
@@ -117,7 +117,7 @@ describe('Automerge.Frontend', () => {
       assert.deepEqual(doc1, {birds: ['chaffinch', 'goldfinch']})
       assert.deepEqual(doc2, {birds: ['goldfinch']})
       assert.deepEqual(req2, {requestType: 'change', actor, seq: 2, deps: {}, ops: [
-        {obj: birds, action: 'del', key: `${actor}:1`}
+        {obj: birds, action: 'del', key: 0}
       ]})
     })
 
@@ -166,41 +166,12 @@ describe('Automerge.Frontend', () => {
         assert.deepEqual(doc1, {counts: [new Frontend.Counter(1)]})
         assert.deepEqual(doc2, {counts: [new Frontend.Counter(3)]})
         assert.deepEqual(req1, {requestType: 'change', actor, seq: 1, deps: {}, ops: [
-          {obj: ROOT_ID, action: 'makeList', key: 'counts',     child: counts},
-          {obj: counts,  action: 'ins',      key: '_head',      elem: 1},
-          {obj: counts,  action: 'set',      key: `${actor}:1`, value: 1, datatype: 'counter'}
+          {obj: ROOT_ID, action: 'makeList', key: 'counts', child: counts},
+          {obj: counts,  action: 'ins',      key: 0},
+          {obj: counts,  action: 'set',      key: 0, value: 1, datatype: 'counter'}
         ]})
         assert.deepEqual(req2, {requestType: 'change', actor, seq: 2, deps: {}, ops: [
-          {obj: counts, action: 'inc', key: `${actor}:1`, value: 2}
-        ]})
-      })
-
-      it('should coalesce assignments and increments', () => {
-        const [doc1, req1] = Frontend.change(Frontend.init(), doc => doc.birds = {})
-        const [doc2, req2] = Frontend.change(doc1, doc => {
-          doc.birds.wrens = new Frontend.Counter(1)
-          doc.birds.wrens.increment(2)
-        })
-        const birds = Frontend.getObjectId(doc2.birds), actor = Frontend.getActorId(doc2)
-        assert.deepEqual(doc1, {birds: {}})
-        assert.deepEqual(doc2, {birds: {wrens: new Frontend.Counter(3)}})
-        assert.deepEqual(req2, {requestType: 'change', actor, seq: 2, deps: {}, ops: [
-          {obj: birds, action: 'set', key: 'wrens', value: 3}
-        ]})
-      })
-
-      it('should coalesce multiple increments', () => {
-        const [doc1, req1] = Frontend.change(Frontend.init(), doc => doc.birds = {wrens: new Frontend.Counter()})
-        const [doc2, req2] = Frontend.change(doc1, doc => {
-          doc.birds.wrens.increment(2)
-          doc.birds.wrens.decrement()
-          doc.birds.wrens.increment(3)
-        })
-        const birds = Frontend.getObjectId(doc2.birds), actor = Frontend.getActorId(doc2)
-        assert.deepEqual(doc1, {birds: {wrens: new Frontend.Counter(0)}})
-        assert.deepEqual(doc2, {birds: {wrens: new Frontend.Counter(4)}})
-        assert.deepEqual(req2, {requestType: 'change', actor, seq: 2, deps: {}, ops: [
-          {obj: birds, action: 'inc', key: 'wrens', value: 4}
+          {obj: counts, action: 'inc', key: 0, value: 2}
         ]})
       })